OnlineRetailPOS-main/requirements.txt
Django==4.1.13
django_admin_rangefilter==0.8.6
django_colorfield==0.7.2
django_import_export==2.8.0
djangorestframework == 3.15.0
import_export==0.2.67.dev6
pandas==2.2.1
numpy == 1.26.4
plotly==5.10.0
python-dotenv==0.21.0
python_dateutil==2.8.2
pytz==2022.1
requests==2.31.0
django-jquery==3.1.0
django-mathfilters==1.0.0
django-admin-logs==1.0.2
python-escpos==2.2.0
# mysqlclient==2.2.0
# psycopg2==2.9.9 
psycopg2-binary==2.9.9 
# sshtunnel==0.4.0
django-tenants
gunicorn==20.1.0

